+++ src/test/java/com/example/event/domain/PhoneNumberTest.java
@@ -1,28 +1,12 @@
 package com.example.event.domain;
 
-import com.example.event.domain.value.Email;
+import static org.assertj.core.api.Assertions.assertThat;
+
 import com.example.event.domain.value.PhoneNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.*;
-
-public class ValueObjectTest {
-
-    @Test
-    @DisplayName("이메일 주소가 같으면 같은 객체로 판단한다.")
-    void emailValidateSuccessfully() {
-
-        //given
-        String Email = "abc@abc.com";
-
-        //when
-        Email email1 = new Email(Email);
-        Email email2 = new Email(Email);
-
-        //then
-        assertThat(email1).isEqualTo(email2);
-    }
+public class PhoneNumberTest {
 
     @Test
     @DisplayName("핸드폰 번호가 같으면 같은 객체로 판단한다.")
