+++ src/test/java/com/example/event/domain/TicketTest.java
@@ -4,6 +4,9 @@ import com.example.event.TicketTestFixtures;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+/*
+    to do
+ */
 public class TicketTest {
